Update ReviewFileController delete action
We need to delete the uploaded file according to the radio button
selection. This commit adds a RemoveFileInput input object, which is
used in the delete action to determine whether to delete the file or not.

I'll refactor the delete and confirmation actions to reduce duplication
in a subsequent PR.