Log more information about S3 operations
It's useful to log the S3 object key when we log operations on files
uploaded for a file upload question so that we can link together
operations on the object during is lifecycle if we ever need to track
what happened to an object.

Additionally, log when a file is deleted from S3 or retrieved to be
attached in an email so we have greater visibility over the object
lifecycle.

Added `FileUploadLogger` to ensure that we always log these operations
with the `s3_object_key` field.