Move all StepFactory usage into Journey
Move all methods in `Context` that interact with the StepFactory into
the Journey class. The means that the Journey now has responsibility
for constructing `Steps` and navigating the form filling journey based
on the saved answers.

`Context` now becomes a wrapper for `FormContext` and `Journey`
methods, tying them together to read from the session for an active
form.

In order to store submission data in the database and send the email
later, we want to be able to construct an instance of `Journey` from
the data stored in the database rather than the session, and this
refecator gets us closer to being able to do this.

Co-authored-by: David Biddle <david.biddle@digital.cabinet-office.gov.uk>
Co-authored-by: Jamie Wilkinson <jamie.wilkinson1@digital.cabinet-office.gov.uk>
Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>