Add class for storing and accessing confirmation details
Add a `ConfirmationDetailsStore` which just deals with storing the
submission confirmation details on the session so that they can be
displayed back to the user on the confirmation page.

We want to separate this from `FormContext` so that this just deals
with the answers. The end goal is to rename this to be `AnswerStore`
which can be populated from either the session or from submission
details we will store in the database.

Co-authored-by: Jamie Wilkinson <jamie.wilkinson1@digital.cabinet-office.gov.uk>
Co-authored-by: David Biddle <david.biddle@digital.cabinet-office.gov.uk>