Rename FormContext -> SessionAnswerStore
The FormContext now only deals with saving and retrieving answers from
the session. Rename this class to SessionAnswerStore to more
accurately reflect its purpose. We will subsequently add a
DatabaseAnswerStore class which will implement the `get_stored_answer`
method. This will allow us to create a `Journey` using either the
SessionAnswerStore or DatabaseAnswerStore.

Co-authored-by: Jamie Wilkinson <jamie.wilkinson1@digital.cabinet-office.gov.uk>
Co-authored-by: David Biddle <david.biddle@digital.cabinet-office.gov.uk>