Add DatabaseAnswerStore to read answers from the database
In order to construct an instance of `Journey` and build the
steps based on the stored answers, it needs an instance of a class
that implements the `get_stored_answer` method.

The new DatabaseAnswerStore class represents answers loaded from the
database, where we will store a hash with the page IDs as keys and the
answers as values.

Add tests to ensure that the Journey works with either a
`SessionAnswerStore` or a `DatabaseAnswerStore`, and a shared example
included in the unit tests for these two classes to enforce the shared
interface.

Co-authored-by: Jamie Wilkinson <jamie.wilkinson1@digital.cabinet-office.gov.uk>
Co-authored-by: David Biddle <david.biddle@digital.cabinet-office.gov.uk>