Pass Journey to AwsSesSubmissionService
We are now able to theoretically construct a Journey from data stored
in the database by using a DatabaseAnswerStore.

Construct the AwsSesSubmissionService with the Journey rather than
the CurrentContext to uncouple sending submission emails from
interacting with the session. This gets us closer to being able to
store the submission data in the database and send emails using the
database record.

This will allow us to:
* Send emails asynchronously from a background process
* Retry bounced emails