Add configuration to run Solid Queue locally
Add configuration that will allow Solid Queue to be run locally. By
default, the Async adapter is configured as the Active Job adapter
locally, which runs jobs in an in-process thread pool. This will run
enqueued jobs, but not recurring tasks which are configured in
`config/recurring.yml` specifically for Solid Queue.

This commit add the database congiguration to connect to the Solid
Queue database when running Rails in the development enviorment
locally. This will only be used if the lines in `development.rb` to
set Solid Queue as the Active Job adapter are uncommented.

If these lines are uncommented, Solid Queue can be run locally by
running `./bin/jobs`.