Add job to delete submissions after 7 days
Add a job that will find all submissions that have been sent and were
updated earlier than 7 days ago and:

- delete any files for file upload questions from S3
- destroy the database record for the submission

If deleting an individual submission fails for any reason, log the
error and send it to Sentry and continue deleting other submissions.

This job will be run as a recurring task on a schedule, but this
commit does not configure this yet.