Attach multiple files with same name
We were storing uploaded files in a hash in aws_ses_submission_service
with the filename as the key before passing this hash into the mailer.
This meant that if there were multiple files with the same filename, the
most recent file would overwrite any previous files with the same name.

This commit replaces the hash with an array of hashes, each of which has
a `name` key and a `file` key. This means that there can be multiple
files with the same name and they'll still be separate hashes.

In the mailer, we can call `attachments[file[:name]]` with multiple
files and they won't overwrite each other. This is because `attachments`
doesn't behave like a standard Ruby hash, it's a
[Mail::AttachmentsList](https://www.rubydoc.info/gems/mail/Mail/AttachmentsList)
with a highly customised `[]=` method.