Raise error if wrong number of files attached
This commit adds a check to see whether the number of files passed into
the mailer matches the number of attachments that will be sent. This
should prevent any issues if an attachment fails or is overwritten in
the Mail gem - for example, if the Mail::AttachmentsList were to stop
supporting multiple files with the same name.