Add temporary rake task to set mode on submissions
When we first started storing submissions, we were not storing the
`mode`, which is required to retrieve the form from forms-api.

This means that the DeleteSubmissionsJob is failing for these
submissions. Add a temporary rake task to set the `mode` on these
submissions.

Add another task to list submissions without a mode so we don't miss
any.

These rake tasks can be removed as soon as these submission records
are fixed.