Add form_document column to submissions table
In order to construct the email from the answers and forms data when
the job to send a submission email is run, or when we are attempting
to resend a bounced submission later on, we need to ensure we are
using the exact version of the form that existed when the form was
submitted.

For now, we will store the JSON blob of the form document on the
submission and use this, rather than retreiving the form from
forms-api. In the future, we might look at implementing versioning for
forms so we do not need to store a snapshot of the form document on
each submission.