Log attributes consistently when performing jobs
Use CurrentAttributes to set attributes that will be logged for all
log lines when a submission is being processed. This means that we
don't need to manually add logging attributes and anything logged by
nested calls gets the same attributes.

Have added a separate CurrentJobLoggingAttributes class to distinguish
from attributes available for HTTP requests.

Log the message_id for the SES message as mail_message_id to be
consistent with what it's called in the database and code, instead
of `submission_email_id` which is how we log it for submissions using
Notify.