Redirect to exit_page if answer matches
Exit pages do not route on to other questions in the form when the
condition is met.

To represent this, Step#next_page_slug_after_routing
returns nil when the page matches.

This allows the journey code to stop looking for the next page and
return. Exit page conditions already have a goto_page_id set to nil, so
we could let this value be passed on more implicitly.

Handling the case early clearly shows our intention. Exit_pages don't go
to another page when they have been selected.