Reduce filename limit to 100 characters
By default the Windows API supports a [maximum file path of 256
characters](https://learn.microsoft.com/en-us/windows/win32/fileio/maximum-file-path-limitation).
This is the entire path, not just the filename as assumed in the commit
where we initially added the truncation.

Because of this, we need to reduce the maximum filename length we allow
so that the file can be downloaded more easily by form processors.

This commit sets a limit of 100 characters, which is still an unusually
long filename but gives some buffer space to allow for the length of the
directory name.