Add exit_page route, controller, view
To show exit pages to form fillers we need to add a route, controller
and view.

Exit pages are represented by conditions with exit_page_markdown set.

For the route, we use the page slug and add /exit to the URL. This works
because pages can only have one exit page at the moment. We might need
to add condition id or make other changes when pages can have more than
one exit page or when the exit page is the secondary route.

The exit page controller inherits from page because it uses the current
form, mode and step.

We also redirect users back when they cannot view the current step
because they haven't completed the form. The exit page is accessible as
soon as the user can view the question step, regardless of the answer to
the selection question.