Return hash for skipped files in show_answer_in_csv
The logic added in #1253 returned `nil` in `show_answer_in_csv` if the
original_filename is blank. `show_answer_in_csv` expects a hash, so this
commit updates it to return a hash in all cases.

This should've been caught by a test but I didn't cover the nil case in
the tests previously - it's now covered.