Remove mode span from visually hidden text
We were using the question_text_with_optional_suffix_inc_mode helper in
the visually hidden text for the check your answers page change link.

We used to use this method in headings to make the mode clear for users
of assistive technology. We now have a different implementation for
that.

As far as I can tell this mode string doesn't offer any benefit in the
Check Your Answers context, and currently the HTML isn't stripped so it
will actually read out incorrect text in screen readers.

This commit removes it and replaces it with the
question_text_with_optional_suffix, which I think was the intended
behaviour anyway.

Since that was the only remaining use of
question_text_with_optional_suffix_inc_mode this commit also removes it
altogether.