Add metric for time to execute send submission job
Add a CloudWatch metric that sends the time after the job to send a
submission was scheduled to be run. The job to send the submission is
scheduled to be run immediately, so under normal circumstances it
should finish executing shortly after this time.

We want to be able to monitor when we are not able to keep up with
processing jobs on the queue. In this case, we would expect this time
to begin rapidly increasing.

We're using the time since the scheduled time, rather than the time
the job was enqueued to account for when a job fails and is retried.
In these cases, this time will be measured from the time when the
retry is scheduled to be run and so won't skew the statistics. We want
a different metric to measure failure counts. If the `scheduled_at`
isn't set for the job, fallback to using the `enqueued_at`. In reality
the `scheduled_at` should be set, but it isn't for tests.

Use the `cloudwatch_metrics_enabled` setting that we use for deciding
whether to send per form metrics to decide whether to send the new
metrics. We can introduce a separate setting in future if we want to
enable the different metrics individually.