Add page heading to check your answers page
In some cases, in particular for file upload, form creators may use
similar question text for multiple pages with different guidance text.

In these cases, we think it would be helpful to the user to have the
page heading on the Check Your Answers page so that they can tell which
question is which.

Because this has implications for existing pages with guidance, we're
limiting this to file upload questions for now.