Rename CloudWatch metrics to match AWS conventions
We're adding additional metrics to monitor the health of our
background jobs, and we reviewed the naming conventions of our
existing metrics as part of this.

We want to follow the AWS metrics naming conventions by making
namespaces, metric names and dimension names all CamelCase.

To do this, we need to start sending new metrics with the desired
names alongside the existing metrics. We can then switch over
forms-admin to read the new metrics once we have been sending metrics
for at least 8 days, as forms-admin only displays metrics for the
previous 7 days. Once forms-admin is reading the new metrics, we can
stop sending the old metrics.

For the new metrics, use the namespace name `Forms`, and add an
`Environment` dimension rather than including this in the namespace.