Add queue prioty order
We want to make sure that the most important jobs are not blocked by
less important jobs running.

The most important job is sending the queue metrics, as we want these
to be sent on a regular schedule so that we know the health of the
queues. Ensure jobs from this queue are run before all other jobs.
After this, just run jobs in the order they were queued.

We don't have any particularly intensive background jobs right now, so
we probably don't need to prioritise the job to send submission emails
above these. We can revisit this if we introduce longer running
background jobs that we'd prefer to defer to running in quieter
periods if we experience high demand for sending submission emails.