Use chromedriver v128 in github actions
We are seeing intermittent failures when feature tests run in github
actions. These failures are related to selenium and chromedriver. They
are false positives and do not mean the code is broken.

This commit copies the solution gov.uk applied to their codebase.

https://github.com/alphagov/publisher/pull/2596/commits/68adc8858782949a1a38f4eb582a17462eb0e51c

This change should only be needed until the chromedriver/chrome version
issue is fixed.