Set sent_at on submission
We were relying on the updated_at to decide when to delete a
Submission. However, we're now updating the Submission record when we
get notification via SNS that a submission has been successfully
delivered. We don't know how long it will take to get and process this
notifaction, and we don't want to delay deleting submission past our 7
day retention period.

Set the sent_at on the submission when we've made the request to SES
to send the submission so we can use this to decide when to delete
submissions.