Use sent_at to get submissions to delete
Use sent_at rather than updated_at when getting submissions to delete.
This does not include a provision to delete submissions that were
created before we started setting the sent_at - but these will only
be test submission so we can either ignore these or manually delete
them.

We also no longer need to check that the `mail_message_id` is set as
the `sent_at` tells us that we've received a success response when we
made the request to SES to send the email.