Remove fallback to get forms from forms-api
For submissions sent via SES, we now store the JSON for the form
document on the Submission. We left a fallback to retrieve the form
from forms-api if this wasn't set so that we could successfully
delete Submissions that already existed at this time. Enough time has
passed for all these Submissions to have been deleted by the
recurring job, so the fallback can be removed.