Send uploaded files with S3 submissions
When the `submission_type` is set to `s3` for a form, and the form has
answered file upload questions, send the uploaded files to the s3
bucket set on the form.

Do this by copying the files to the target bucket and then deleting
them from our S3 bucket. Do not copy the object tagging when we do the
copy as the form processor does not need the tag added by GuardDuty
and to copy the tagging we would need to ask form processors to allow
the s3:PutObjectTagging permission for our IAM role.

Copy the files before uploading the submission S3 to the target bucket
so that form processors are able to add automations on the CSV file
being added and be sure that any uploaded files referenced in the CSV
are already present.