Check that number of files is correct
Prior to this commit, if duplicate filenames made it past our suffixing
logic somehow we wouldn't know about it - the last file with the
duplicate filename would overwrite the previous ones in the
uploaded_files_in_answers hash, the email would be sent and only the
last file with a given filename would be attached to the email.

This commit adds logic so that if the uploaded_files_in_answers hash
contains a number of files that's different from the number of completed
file questions, the submission job will throw an error. This should
catch duplicate file names and any other issue that causes the wrong
number of files to be passed to the mailer.