Omit submission reference in filenames for S3 submissions
For sending uploaded files to S3, we don't want to include the
submission reference in the filename as this will already be in the
folder name part of the object key.

Pass in a parameter to CsvGenerator#write_submission to specify
whether or not the CSV is for an S3 submission or not. Pass this
parameter into the `show_answer_in_csv` method on the question and use
this in file questions to decide whether or not to include the
submission reference.