Enable none of the above condition triggers
This bypasses a quirk with how we stored our condition answer values for "None of the above" answers. Intead of storing the answer text, we've been storing `none_of_the_above`, which has meant that the answer value check used during conditional routing hasn't been working correctly.

This fixes it so that it will work when the answer value for a condition is `none_of_the_above`, and the answer given to the selection list question is "None of the above".