Add file type and size to logging attributes
Add an `answer_metadata` logging attribute to log lines, and populate
this for `file` questions with the file_type and file_size attributes
for the answer. We can then consume our standard request logs for the
request to submit a file upload answer rather than needing to add
explicit logging for file upload in order to extract and report on
this data.