Fix content on goto question routing error page if route is secondary skip
For a secondary skip route the question number shown when the goto page
is before the start of the route was wrong, because we consider
secondary skip routes to belong to the branch question.

This commit changes the logic for the check_goto_page_routing_error so
that it uses the question number for the route's check page rather than
the route’s routing page, and then it will work for both skip and
secondary skip routes.