Refactor Mode
By convention, and as enforced in our routes configuration, for a live
form the mode string should be "form"; however we don't represent this
anywhere in the `Mode` model.

This commit makes some changes to to do this:

- Change default mode string to "form"
- Check if the mode string is "form" in live? before checking !preview?
- Implement the to_param method from ActiveModel::Conversion to return
  "form" as the url parameter if the form is live

These should make it easier to write tests with the Mode model as well
as make the intention of the mode string clearer.