Change FormHeaderComponent to use service navigation component
Version 5.6.0 of the Design System includes a new [Service
navigation][1] component. And in the newest release GOV.UK Frontend
v5.9.0 the service name in the top navigation will be retired.

This commit moves the form name and start page URL into the service
navigation component as the service name and service URL, and also
changes the header border to be full width to match the requirements of
the service navigation's design [2].

[1]: https://design-system.service.gov.uk/components/service-navigation/
[2]: https://design-system.service.gov.uk/components/service-navigation/#change-the-blue-colour-bar-under-the-govuk-header-to-full-width