Stop sending old form started/submitted metrics
We've added new metrics that have a different name, namespace and
dimensions for displaying the form metrics to form creators and were
sending these alongside the old metrics. The new metrics better
conformed to the CloudWatch naming conventions. Forms-admin is now
using the new metrics, so we can stop sending the old metrics.