Add defaults for GOVUK Design System Form Builder
In order to support Welsh translations, add configuration for the
govuk_design_system_formbuilder gem to set default content for the
components we can configure it for.

The documentation for govuk_design_system_formbuilder suggests adding
this in a Rails initialiser file. However, if we try to set this
configuration during initialization using translations, the
localization has not yet been initialized so loading the translations
fails. Due to this, we are setting the configuration in an
after_initialize callback.

If translations were to fail to load, this would break our feature
tests - so these test we are doing the initialization in the correct
place.