Fix email confirmation link to call charges
Currently the form-filler email confirmation includes a link to the call
charges page when the form has a contact phone number set.

This link has text but no URL. This is because the email confirmation
uses an incorrect attribute for the support_details.

It seems the attribute was always wrong and the tests written to expect
a link with no URL.

This commit changes the attribute to the correct value and updates the
tests.