Display submitted at time in local time
Emails sent using SES were displaying the submitted at time in UTC
rather than in local time. This is because we are using the created_at
time for the submission when processing the SendSubmissionJob, which
is stored as UTC.

We need to convert the UTC time to the time we've configured to
display submissions in, which is London time.