Add fix for test failure caused by render blocking
This test was failing because of a race condition - which works
something like this:

- the beforeEach callback uses innerHTML to add HTML to the document,
  which then queues a task to parse and render the new HTML.
- the next step in the beforeEach callback adds the event listener to
  details events (this is run synchronously before the parse and render
  tasks are picked up from the queue)

Because of this, the event listener isn't actually added to the details
element, since it hasn't been parsed yet.

This commit fixes the test by adding a 0ms timeout in between adding the
HTML and running the event listener. This adds a task which does nothing
to the end of the task queue, and waits for it to complete before
proceeding. So the event listener won't be added until all of the tasks
on the queue (including the parse and render) have been completed.