Override default content in design system components
Configuring default values for the GOVUKDesignSystemFormBuilder
doesn't work for localisation, as we need to pick the language per
request - but these translations were evaluated at the point the
application starts.

Instead, pass in translations to all the components that were using
the default values rather than allowing them to use the default
content.