Set locale for routes not in form scope
For the Cookies and Accessibility statement page, we do not have a
form associated with the route to get the language from.

So we can show these pages in Welsh, accept a `language` query
parameter to use for setting the locale. If given an unsupported
locale, set the locale to "en".