Stop setting mail_status to delivered
It is possible to receive a bounce and delivery notification for the
same SES message. This means that if a delivery notification is
processed after a bounce notification for the same submission, we will
mark that submission as delivered incorrectly, and it could go missing.
This hasn't happened yet, but could.