Include language query param in links
If the current locale is a locale other than English, include a
language query parameter in the footer links for the Accessibility
statement and the Cookies page so that these are rendered with
translations for the current locale.

Accomplish this by adding a ViewComponent for the footer that includes
this logic. We're adding a view.html.erb file for this component as
soon we'll also need to pass in a block of HTML for the licence
section of the footer as this needs to include the localized licence
statement with a link.