Add code from GOV.UK Notify for validating email addresses
We've been seeing errors in production from the GOV.UK Notify API
from invalid email addresses:

```
ValidationError: email_address Not a valid email address (Notifications::Client::BadRequestError)
```

This is happening because when we ask form fillers for an email address
for their confirmation email, the validation we do to check whether the
email address is valid is much simpler than the validation done by
Notify themselves.

It's preferable that we pre-validate the emails on our servers rather
than rely on Notify to validate for us, especially if we want to start
sending confirmation emails asynchronously.

So we need to have the same validation logic for email addresses as
Notify. Unfortunately the rules around what is a valid email address are
quite complicated.

This commit copies a bunch of code from the alphagov/notifications-utils
repo related to checking if an email address is valid. That code is
written in Python, so the code in this commit has been lightly
rewritten, but the emphasis was on keeping similarity with the original
rather than making it fully idiomatic Ruby.

This approach means that hopefully it is easier to stay in sync with
Notify if they ever need to change their email address validation.

Note that to support this code, an IDNA2008 library was needed to match
the [Python idna library] used in the Notify code. I chose the [uri-idna
gem], because it appears to be the best supported currently, and it does
not rely on external C dependencies.

[Python idna library]: https://pypi.org/project/idna/
[uri-idna gem]: https://rubygems.org/gems/uri-idna