Set locale when handling not found errors
It seems that the locale set using `around_action` in the controller
has been reset when the block for `rescue_from` is run.

Set the locale using the language attribute for the form, if it is
present. Access the form using the `@current_form` instance variable
rather than the `current_form` method to avoid re-attempting to
retrieve the form from forms-api if the error is due to a form not
existing.