Send confirmation emails when config missing
Previously confirmation emails were never sent when the "what happens next" text or support details were missing for a form. This created confusion for form creators when submitting of draft version of forms. This now allows sending of confirmation emails, but instead with placeholder text in the email informing the form creator that the configuration needs to be added.

This placeholder text should never be visible on a live form as "what happens next" and support contact details are required to make the form live.