Add find_archived method to Api::V1::FormSnapshotRepository
This commit adds a find_archived method to
Api::V1::FormSnapshotRepository, to retrieve an archived snapshot of a
form. I've added this method instead of passing a mode into
find_with_mode, as I want to be able to retrieve an archived form when
in the live mode, and this felt clearer about intent.