Move formatting support contact details to Mailer
This functionality is only needed by the submission confirmation mailer
and not by other parts of the form submission service. There's probably
more refactoring that could be done, but this move the concerns into a
more logical class.

This work is done to make it easier to refactor the
FormSubmissionsService.