Render the form archived page for archived forms
This commits amends the `check_available` before action to attempt to
retrieve an archived form if retrieval of a live version failed. If an
archived form is found, we render the form archived page with a 410
gone status code.

Note the use of `prepend_before_action`. This is because the
`set_request_logging_attributes` before_action defined in the parent
controller was running first, which itself access the current form.
This raises an ActiveResource::ResourceNotFound error in the case of a
form being archived, and so we need to run `check_available` first.