Base64 encode file attachments
By default ActionMailer tries to encode any file attachments in Base64.

We've seen a problem in production where this encoding fails with an
error 'invalid byte sequence in UTF-8'. We think this happens when a
binary file has been uploaded with a text file extension, so the
ActionMailer logic expects a text file but receives binary characters.

To fix this, we can encode the files in Base64 ourselves before passing
them to ActionMailer, which bypasses the failing logic. This allows us
to pass these disguised binary files to ActionMailer without errors.
Other files should be unaffected and the resulting emails should be
otherwise identical.