Use new column names for Submissions
This updates the code to start using delivery_status and last_delivery_attempt columns instead of mail_status and sent_at. The code provides support to use data in the old columns too. This is part of the data migration to rename these columns. Future changes will migrate add data to the new columns and drop the old ones.