Stop stripping whitespace in email validation
We currently strip whitespace in the email validation before making
other checks on the format.

This validation doesn't modify the original attribute, which means that an
email with surrounding whitespace will be considered valid.

Passing an email with line breaks on to the mailer causes an exception
to be raised.

To ensure emails are valid, we stop stripping the whitespace. This means
the caller is responsible for removing it before validation.

This makes the validation slightly clearer and is less likely to
lead you to think the attribute you are validating will be modified.