Add rake task to migrate data from old columns in Submissions
"delivery_status" is to replace "mail_status" and "last_delivery_attempt" is to replace "sent_at". To drop the old column, this adds a rake task to ensure that data iin the new columns is in sync with old columns. After this task is run, we can remove use and drop the old columns.