Update Gemfile to fix deprecation warnings
The latest version of Bundler was complaining that

    Platform :mingw, :x64_mingw is deprecated. Please use platform :windows instead.

This commit updates the line for the debug gem in to oblige it. This
should work with all installations of this app, since Bundler will
always install the version of Bundler specified in the lockfile before
proceeding.

We also add `require: "debug/prelude"` to the line for the debug gem to
match what is generated by `rails new` (see [[1]]); this might also give
a small speedup.

[1]: https://github.com/rails/rails/commit/7254549dcd95a6b9f71dab48b2c0a56956682e9a