Reapply "Merge pull request #1540 from alphagov/rebrand-ses-email-template"
This reverts commit 36b80f3cdbc2549ae8edafe94151a86e02c139d7.

We previously implemented changes to the submission email template, but
reverted these because of [an incident](https://docs.google.com/document/d/193BkGQZjyqCgNEbTe8MCF9Xpj8EQuQzDa3KD-1CBKpU/edit?tab=t.0).

This commit readds those changes so that we can add a fix for the issue
that caused the incident.