Add missing Outlook conditional comment
Adds a missing closing tag for an Outlook conditional comment. We had an
`if !mso` conditional comment targeting non-Outlook clients. By failing
to close this, we accidentally marked all subsequent HTML in the email
as being for non-Outlook clients only. This meant that in classic
Outlook, all of this content was ignored and not displayed, including
all of the user data in the email.