Improve formatting
The formatting in this file was wrong, likely because automatic
formatters don't cope well with Outlook conditional comments. We don't
currently enforce any automatic formatting on .html.erb files, but
individual devs may have formatters enabled in their IDEs. In this case
I believe the formatting issue was caused by my local `html-beautifier`.

This commit makes some manual formatting to make it easier to understand
the structure of this layout.