Add helper to redact email addresses from an error message
We want to send the error message we get when ActionMailer is not able
to parse the "To" email address for a Mail message to Sentry so we can
try to figure out why certain invalid email addresses are getting
through our validation.

However, we do not want to sent PII to Sentry. This helper method
attempts to obfuscate email addresses, retaining any special
characters which will hopefully give us enough information while
ensuring the email address is not identifiable. We replace the @ with
(at) so that Sentry doesn't match an email address in the
action_mailer_error we send and strip it out completely.