Raise error when there is an ActionMailer error for To email address
Previously, when ActionMailer failed to parse the To email address for
the confirmation email, an error would not be raised until the
GovukNotifyRails::Delivery code attempts to call `message.to.map`
which raises an error due to `message.to` returning a String rather
than a Map.

It returns a String rather than a map only when there is an error
parsing the "To" field.

This commit makes it so that we look at the `mail.messages.errors` to
establish whether there is an error with the "To" field. If there is,
we send an error to Sentry and raise a ConfirmationEmailToAddressError
to be handled by the CheckYourAnswersController in a subsequent
commit.